Merge solutions if they are found by #cannot_be.
While reducing possible solutions it is quite possible that the last
option is the correct one. In that case, that solution should be
included in the return Hash.

Also realized that the Hash return by resolve! cannot be used to set all
values, since it will only contain the identifiers that was resolved
from that specific call. Since there is bound to be identifiers that are
resolved from the initial setup, a simple call to #scan and #resolve
will not give all identifiers, since it will miss the first.