Major restructuring.
Extracted the resolving parts into SolutionSpace to distinguish the
puzzle and the state of the resolving functionality.
It is now possible to step through the resolution since resolve! with
yield before each step.
Extracted the possible values an identifier can have into Possibilities.
Constraint now inherits from Array, since it in essence is just a set of
identifiers.